pyproject.toml
[tool.black]
line-length = 79
target-version = ['py37']

[tool.pytest.ini_options]
minversion = "6.0"
testpaths = [
    "tests",
]
python_files = "tests.py test_*.py *_tests.py"
addopts = "--strict-markers --showlocals -n auto --dist loadscope"

[tool.tox]
legacy_tox_ini = """
[tox]
isolated_build = True
envlist = py37, py38, py39

[gh-actions]
python =
    3.7: py37
    3.8: py38
    3.9: py39

[testenv]
whitelist_externals = poetry
commands =
    poetry install -v
    poetry run pytest --cov=tests/ --cov=panimg/
"""

[tool.poetry]
name = "panimg"
version = "0.2.0"
description = "Conversion of medical images to MHA and TIFF."
license = "Apache-2.0"
authors = ["James Meakin <12661555+jmsmkn@users.noreply.github.com>"]
readme = "README.md"
repository = "https://github.com/DIAGNijmegen/rse-panimg"
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Intended Audience :: Healthcare Industry",
    "Operating System :: POSIX",
]

[tool.poetry.dependencies]
python = "^3.7"
pydantic = "*"
numpy = "*"
SimpleITK = "*"
pydicom = "*"
Pillow = "*"
openslide-python = "*"
pyvips = "*"
tifffile = "*"

[tool.poetry.dev-dependencies]
pytest = "*"
pytest-xdist = "*"
pytest-randomly = "*"
pytest-cov = "*"
tox = "*"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
